requirements/requirements.txt
django==1.8
django-material==0.5
djangorestframework
drf-nested-routers